pyproject.toml
[project]
name = "ai-knowledge-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "pytest>=9.0.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
